initrdscripts: Expand the data partition up to the reserver part of the disk
The reserved part of the disk starts at sector 58724353.
The Nvidia bootloader expectes these two binaries in specific locations on the disk.
These locations are hardcoded in the ROM bootloader and cannot be changed. These locations
are at sector 58724353 and 58746533 respectively. We will place these binaries in those
locations instead of the current location.

Signed-off-by: Theodor Gherzan <theodor@resin.io>