linux-tegra: Fix modules not loaded
With scmversion enabled module loading
fails because the revision length doesn't
match between the kernel and modules dir.

We're not interested in having commit
SHAs in the kernel version anyway, so
let's disable this, like we do for the
rest of the jetson boards.

Changelog-entry: linux-tegra: Fix modules not loaded
Signed-off-by: Alexandru Costache <alexandru@balena.io>