modemmanager: Don't include u-blox mode switch script for now
The u-blox mode switch fails frequently, and is
a known issue, addressed in meta-balena ticket 2039.
The script can also potentially trigger a reboot right
after the board boots in the new OS. Let's not include this
script until a proper fix is in place, because the SKX2
board uses this type of modem.

Changelog-entry: modemmanager: Don't include u-blox mode switch script for now
Signed-off-by: Alexandru Costache <alexandru@balena.io>